requirements.txt
Django==1.11.7
dj-database-url==0.4.2
dj-static==0.0.6
django-appconf==1.0.2
django-compressor==2.2
djangorestframework==3.7.3
drf-nested-routers==0.90.0
gunicorn==19.7
six==1.11.0
static3==0.7.0
wsgiref==0.1.2
django_extensions
jsonfield
